SQL-2249: test derive_schema_for_views (#780)
* update partialeq impl for jaccardindex

* SQL-2175: integration test for derive_schema_for_partitions?

* refactor for local clippy warnings

* update for data loader changes

* refactor tests for commonality

* cleanup

* add view schemas to consts, use default

* SQL-2249: test derive schema for views

* combine similar tests

* remove file

* data loader updates

* assuage clippy's concerns

---------
